requirements.txt
# GlassBox AI Dependencies

# MCP Protocol (Model Context Protocol)
mcp

# OpenAI API
openai>=1.54.0

# Future: Multi-model support
# anthropic>=0.40.0
# google-generativeai>=0.3.0

# Database (SQLite is built-in to Python)

# Agent v2
pydantic>=2.0.0
pyyaml>=6.0

# Development
python-dotenv>=1.0.0
pytest>=8.0.0

# Dev-only (requires Python 3.12+, install separately):
# pip install mcp-hmr
